Error handling.

+++ application.irq/fpga_drv.c
@@ -39,7 +39,7 @@
                        // Can be found by (GIC IRQ is 121 for this example):
                        //   grep 121 /sys/kernel/irq/*/hwirq
 
-#define FPGA_BASE    0xa0000000
+#define FPGA_BASE    0xa0000000    // Base address of the HW accelerator
 #endif
 
 #define REG_BASE    0x00000000
@@ -110,7 +110,6 @@ static irqreturn_t fpga_int_handler(int irq, void *lp)
  * Driver access methods
  */
 
-/* Driver access routines */
 static int fpga_open1 (struct inode *inode, struct file *file) {
    return 0;
 }
@@ -176,11 +175,11 @@ static long fpga_ioctl1(struct file *file, unsigned int cmd, unsigned long arg){
    {
       case 0:
          //read
-         if(!access_ok((unsigned int *)arg, sizeof(int)))
+         if(!access_ok((unsigned int *)arg, sizeof(long)))
             return -EFAULT;
 
-	 value = readl((volatile unsigned int *)&l.reg_ptr[offset]);
-	 put_user(value, (unsigned long*)arg);
+         value = readl((volatile unsigned int *)&l.reg_ptr[offset]);
+         put_user(value, (unsigned long*)arg);
 
 #ifdef DEBUG
          printk("fpga_drv: Read value %08lx\n", value);
@@ -191,7 +190,7 @@ static long fpga_ioctl1(struct file *file, unsigned int cmd, unsigned long arg){
          //write
          access_addr = l.reg_ptr + offset;
 
-         if(!access_ok((unsigned int *)arg, sizeof(int)))
+         if(!access_ok((unsigned int *)arg, sizeof(long)))
             return -EFAULT;
 
          get_user(value, (unsigned long *)arg);
@@ -214,7 +213,7 @@ static long fpga_ioctl1(struct file *file, unsigned int cmd, unsigned long arg){
 
 /* define which file operations are supported by the driver */
 struct file_operations fpga_fops = {
-   .owner=   THIS_MODULE,
+   .owner   = THIS_MODULE,
    .llseek  = NULL,
    .read    = fpga_read1,
    .write   = fpga_write1,
